refactor: extract _PurpleAgentBase for CAR + entropic adapters — fixes #40
Pulls the shared A2A scaffolding (tool schema conversion, assistant-msg
normalization, session state, tool-call id reconciliation, harness update)
into clawloop/environments/_purple_base.py. CAR and entropic adapters now
override only the two bench-specific seams: _build_initial_messages and
_format_a2a_response.

No behavior change. 600 lines deleted, 379 added across the three files.
